[REFACTOR] 이력서 컨트롤러, view 수정
- 연관되어 있는 모든 이력서 url에 member 제거함
