--- src/main/java/com/project/careerfair/controller/resume/ResumeTestController.java
@@ -1,5 +1,0 @@
-package com.project.careerfair.controller.resume;
-
-public class ResumeTestController {
-    // 이력서
-}
