+++ src/main/webapp/WEB-INF/views/resume/list.jsp
@@ -37,7 +37,7 @@
                   </div>
 
                   <div class="btnBx">
-                      <a id="resumeWrittenBtn" href="/member/resume/write" class="btn btnBl mtuSpImgBefore wrBtn">
+                      <a id="resumeWrittenBtn" href="/resume/write" class="btn btnBl mtuSpImgBefore wrBtn">
                           <span>이력서 등록</span>
                       </a>
                   </div>
@@ -55,7 +55,7 @@
                               <li class="chk">
                                   <div class="col col01">
                                       <div class="tit">
-                                          <a href="/member/resume/${resume.resumeId}">${resume.title}</a>
+                                          <a href="/resume/${resume.resumeId}">${resume.title}</a>
                                       </div>
 <%--                                      <div class="date">2023-553215</div>--%>
                                       <div class="date">
@@ -77,7 +77,7 @@
                                   </div>
                                   <div class="col col02">
 
-                                      <div class="btnCell"><a class="btn btnTopBooth" href="/member/resume/${resume.resumeId}/update">수정</a></div>
+                                      <div class="btnCell"><a class="btn btnTopBooth" href="/resume/${resume.resumeId}/update">수정</a></div>
 
                                       <div class="btnCell"><a class="btn deleteBtn" id="delete_${resume.resumeId}" href="#">삭제</a></div>
 
@@ -97,6 +97,6 @@
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.4/jquery.min.js" integrity="sha512-pumBsjNRGGqkPzKHndZMaAG+bir374sORyzM3uulLV14lN5LyykqNk8eEeUlUkB3U0M4FApyaHraT65ihJhDpQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-    <script src="/js/member/resume/list.js"></script>
+    <script src="/js/resume/list.js"></script>
 </body>
 </html>
