+++ src/main/webapp/js/resume/scout.js
